backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.103.1
uvicorn[standard]>=0.23.2
python-multipart>=0.0.6
pydantic>=2.4.2

# Audio processing
torch>=2.0.0
numpy>=1.24.0
transformers>=4.34.0
librosa>=0.10.0
pydub>=0.25.1

# Networking
requests>=2.31.0

# Environment and logging
python-dotenv>=1.0.0

# Add any additional dependencies your vector_store implementation requires